requirements.txt
MouseInfo==0.0.4
numpy==1.16.4
Pillow==6.1.0
PyAutoGUI==0.9.46
PyGetWindow==0.0.6
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
PyTweening==1.0.3
